datatype: implement MPIR_Typerep_op with dataloop
The fallback should work for dataloop engine.
