+++ src/mpi/datatype/typerep/src/Makefile.mk
@@ -5,7 +5,8 @@
 
 mpi_core_sources += \
    src/mpi/datatype/typerep/src/typerep_ext32.c \
-   src/mpi/datatype/typerep/src/typerep_flatten.c
+   src/mpi/datatype/typerep/src/typerep_flatten.c \
+   src/mpi/datatype/typerep/src/typerep_op.c
 
 if BUILD_YAKSA_ENGINE
 mpi_core_sources += \
